Refactored benchmarks to use criterion + Removed feature test which is unstable + Added run configuration for benchmarks
