+++ .run/Run.run.xml
@@ -1,5 +1,6 @@
 <component name="ProjectRunConfigurationManager">
   <configuration default="false" name="Run" type="CargoCommandRunConfiguration" factoryName="Cargo Command">
+    <option name="buildProfileId" value="dev" />
     <option name="command" value="run --bin reactive-graph" />
     <option name="workingDirectory" value="file://$PROJECT_DIR$" />
     <envs>
