+++ Cargo.toml
@@ -100,6 +100,7 @@ chrono = { version = "0.4", features = ["serde"] }
 colored = "2.0"
 config = "0.14.0"
 convert_case = "0.6"
+criterion = { version = "0.5", features = ["html_reports"] }
 crossbeam = "0.8"
 cynic = { version = "3.7", features = ["http-reqwest"], default-features = false }
 darling = "0.20"
