+++ crates/reactive/model/impl/src/lib.rs
@@ -1,5 +1,3 @@
-#![feature(test)]
-
 pub use entities::*;
 pub use flows::*;
 pub use frp::*;
