+++ crates/reactive/service/impl/Cargo.toml
@@ -31,6 +31,7 @@ reactive-graph-lifecycle = { version = "0.10.0", path = "../../../lifecycle" }
 reactive-graph-type-system-api = { version = "0.10.0", path = "../../../type-system/api" }
 
 [dev-dependencies]
+criterion = { workspace = true, features = ["html_reports"] }
 default-test = { workspace = true }
 log4rs = { workspace = true, features = ["console_appender"] }
 
@@ -43,3 +44,7 @@ reactive-graph-behaviour-service-impl = { version = "0.10.0", path = "../../../b
 
 [lib]
 crate-type = ["lib"]
+
+[[bench]]
+name = "reactive_entity_manager"
+harness = false
