+++ crates/reactive/service/impl/src/lib.rs
@@ -1,5 +1,3 @@
-#![feature(test)]
-
 pub use reactive_entity_manager_impl::*;
 pub use reactive_flow_manager_impl::*;
 pub use reactive_instance_event_manager_impl::*;
@@ -11,6 +9,3 @@ pub mod reactive_flow_manager_impl;
 pub mod reactive_instance_event_manager_impl;
 pub mod reactive_relation_manager_impl;
 pub mod reactive_system_impl;
-
-// #[cfg(test)]
-// mod test;
